backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiogram>=3.0.0
openpyxl>=3.1.0
qrcode[pil]>=7.4.0
pillow>=10.0.0
httpx>=0.25.0
pyjwt>=2.8.0
